+++ ts/components/Button.tsx
@@ -19,23 +19,48 @@ export type ButtonVariant =
 export type ButtonSize = `xs` | `sm` | `md` | `lg` | `xl`;
 export type ButtonType = `button` | `submit` | `reset`;
 
-interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
+export interface ButtonStyleProps {
   /** Adds the active style to the button. */
   active?: boolean;
-  /** Content for the button. */
-  children?: string | JSX.Element[] | JSX.Element;
-  /** Adds class names to the button. */
+  /** Additional classes to apply to the button */
   className?: string;
+  /** Adds disabled style to the button */
+  disabled?: boolean;
   /** Adds the focus style to the button. */
   focus?: boolean;
   /** Controls the size of the button. */
   size?: ButtonSize;
   /** Adds the square style to the button. */
   square?: boolean;
-  /** Controls the type of the button. */
-  type?: ButtonType;
   /** Controls the variant styling of the button. */
   variant?: ButtonVariant;
+}
+
+interface ButtonProps extends ButtonStyleProps, ButtonHTMLAttributes<HTMLButtonElement> {
+  /** Controls the type of the button. */
+  type?: ButtonType;
+}
+
+export function buttonClassNames({
+  active,
+  className,
+  disabled,
+  focus,
+  size,
+  square,
+  variant,
+}: ButtonStyleProps) {
+  return classNames(
+    `h-btn h-btn--${size}`,
+    className,
+    { [`h-btn--${variant}`]: variant },
+    {
+      'h-btn--active': active,
+      'h-btn--focus': focus,
+      'h-btn--square': square,
+      'h-btn--disabled': disabled,
+    },
+  );
 }
 
 const Button = (
@@ -43,6 +68,7 @@ const Button = (
     active = false,
     children,
     className,
+    disabled = false,
     focus = false,
     size = `md`,
     square = false,
@@ -56,16 +82,8 @@ const Button = (
     <button
       ref={ref}
       type={type}
-      className={classNames(
-        `h-btn h-btn--${size}`,
-        { [`h-btn--${variant}`]: variant },
-        {
-          'h-btn--active': active,
-          'h-btn--focus': focus,
-          'h-btn--square': square,
-        },
-        className,
-      )}
+      className={buttonClassNames({ active, className, disabled, focus, size, square, variant })}
+      disabled={disabled}
       {...props}
     >
       {children}