+++ ts/components/Dialog/SimpleAlertDialog.tsx
@@ -1,20 +1,12 @@
 import React, { useRef } from 'react';
-import { AlertDialog } from './AlertDialog';
+import { AlertDialog, AlertDialogProps } from './AlertDialog';
 import { DialogFooter } from './DialogFooter';
 import { DialogHeader } from './DialogHeader';
 import { DialogBody } from './DialogBody';
 import { ButtonSize, ButtonVariant } from '../Button';
 import { DialogFooterActions } from './DialogFooterActions';
 
-interface SimpleAlertDialogProps {
-  /** Content for the dialog. */
-  children: string | JSX.Element[] | JSX.Element;
-  /** Callback function when the dialog is closed. */
-  closeDialog: () => void;
-  /** Controls whether the dialog is open or not. */
-  open: boolean;
-  /** Adds class names to the backdrop element. */
-  backdropClassName?: string;
+interface SimpleAlertDialogProps extends AlertDialogProps {
   /** Controls the size of the footer action buttons. */
   buttonSize?: ButtonSize;
   /** Text for the cancel button. Setting this prop will show the cancel button. */
@@ -27,8 +19,6 @@ interface SimpleAlertDialogProps {
   confirm?: string;
   /** Controls the variant of the confirm button. */
   confirmVariant?: ButtonVariant;
-  /** Adds class names to the dialog wrapper element. */
-  dialogClassName?: string;
   /** Controls the dialog header text. */
   header?: string;
   /** Adds class names to the dialog header element. */
@@ -37,12 +27,6 @@ interface SimpleAlertDialogProps {
   onCancel?: () => void;
   /** Callback function when the confirm button is clicked. */
   onConfirm?: () => void;
-  /** Position of the dialog. Can be "top", "right", "bottom", or "left". Default position is "center". */
-  position?: string;
-  /** Element to focus when the dialog is closed. */
-  returnFocusEl?: string | HTMLElement | (() => HTMLElement) | undefined;
-  /** Controls the size of the dialog. Default size is "medium". */
-  size?: string;
 }
 
 export const SimpleAlertDialog = ({
@@ -58,12 +42,14 @@ export const SimpleAlertDialog = ({
   dialogClassName,
   header = ``,
   headerClassName,
+  id,
   onCancel = () => {},
   onConfirm = () => {},
   open,
   position,
   returnFocusEl,
   size,
+  wrapperClassName,
 }: SimpleAlertDialogProps) => {
   const leastDestructiveRef = useRef(null);
 
@@ -74,11 +60,13 @@ export const SimpleAlertDialog = ({
         backdropClassName,
         closeDialog,
         dialogClassName,
+        id,
         leastDestructiveRef,
         open,
         position,
         returnFocusEl,
         size,
+        wrapperClassName,
       }}
     >
       <>